requirements.txt
jupyter
num2words
opencv-python
pandas
pdf2image
pytesseract
pytorch
sacrebleu
sentencepiece
transformers
